pi/requirements.txt
websockets==12.0
RPi.GPIO==0.7.1
smbus2==0.4.3
picamera2==0.3.17
Pillow==10.3.0
